luna-platform/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
email-validator==2.2.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20
pandas==2.2.3
numpy==2.1.3

# PostgreSQL driver (optional; SQLite works without it)
# psycopg2-binary==2.9.10
